add notifications to job and app endpoints

+++ src/main/java/com/app/mystore/service/ApplicationService.java
@@ -2,9 +2,11 @@ package com.app.mystore.service;
 
 import com.app.mystore.dao.ApplicationDao;
 import com.app.mystore.dao.NotificationsDao;
+import com.app.mystore.dao.UserDao;
 import com.app.mystore.dto.Application;
 import com.app.mystore.dto.Notification;
 import com.app.mystore.enums.NotificationTypeEnum;
+import com.app.mystore.utils.MystoreHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -18,6 +20,15 @@ public class ApplicationService {
     @Autowired
     private NotificationsDao notificationDao;
 
+    @Autowired
+    public UserDao userDao;
+
+    @Autowired
+    public MystoreHelper helper;
+
+    @Autowired
+    public LoginControllerService loginControllerService;
+
     public List<Application> fetchAll() {
         List<Application> applications = dao.fetchAll();
         return applications;
@@ -28,10 +39,6 @@ public class ApplicationService {
         return application;
     }
 
-    public Application fetchByUserID(int userID) {
-        Application application = dao.getByUserID(userID);
-        return application;
-    }
 
     public Boolean updateApplication(Application updateApplication) {
         int result = dao.updateApplication(updateApplication);
@@ -46,7 +53,13 @@ public class ApplicationService {
     public Boolean addApplication(Application newApplication) {
         int result = dao.insertApplication(newApplication);
         if(result > 0){
-            int managerUserID = 6;
+            int managerUserID = 0;
+            try {
+                managerUserID = userDao.getManagerId();
+            } catch (Exception e) {
+                e.printStackTrace();
+                return false;
+            }
             String applicationNotification = "New application from " + newApplication.getFirstName();
             Notification notification = new Notification();
             NotificationTypeEnum notificationTypeEnum = NotificationTypeEnum.JOB_NOTIFICATION;
@@ -68,12 +81,66 @@ public class ApplicationService {
     }
 
     public Boolean deleteApplication(int applicationID) {
-        int result = dao.deleteApplication(applicationID);
-        if(result > 0){
+        Application application = dao.getByApplicationID(applicationID);
+        if (application != null) {
+            String applicationNotification = "Thank you for your application. Unfortunately you were not selected but please apply again.";
+            Notification notification = new Notification();
+            NotificationTypeEnum notificationTypeEnum = NotificationTypeEnum.JOB_NOTIFICATION;
+            notification.setUserId(application.getUserID());
+            notification.setNotification(applicationNotification);
+            notification.setNotificationType(notificationTypeEnum.getType());
+            try{
+                notificationDao.createNotification(notification);
+            }
+            catch (Exception e) {
+                e.printStackTrace();
+                return false;
+            }
+            int result = dao.deleteApplication(applicationID);
+            if(result > 0){
+                System.out.println("here");
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    public Boolean acceptApplication(int applicationID) {
+        Application application = dao.getByApplicationID(applicationID);
+        int row =0;
+        System.out.println(application.getUserID());
+
+        if(application != null){
+            String body = "Hello, We would like to inform you we have accepted your application! Please sign into myStore to view the employee page.";
+            helper.sendEmail(application.getEmail(), body, "Approved application");
+            try {
+                row = loginControllerService.updateRole(application.getUserID());
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            dao.deleteApplication(application.getApplicationID());
+            String applicationNotification = "Congrats on your new job at myStore!";
+            Notification notification = new Notification();
+            NotificationTypeEnum notificationTypeEnum = NotificationTypeEnum.JOB_NOTIFICATION;
+            notification.setUserId(application.getUserID());
+            notification.setNotification(applicationNotification);
+            notification.setNotificationType(notificationTypeEnum.getType());
+            try{
+                notificationDao.createNotification(notification);
+            }
+            catch (Exception e) {
+                e.printStackTrace();
+                return false;
+            }
+        }
+        System.out.println("row  result: " + row);
+        if(row > 0){
             return true;
         }
         else{
             return false;
         }
     }
+
 }
